requirements.txt
streamlit>=1.29.0
PyMuPDF>=1.24.0
Pillow>=10.1.0
numpy>=1.26.0
scipy>=1.11.0
